requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
Cartopy==0.21.1
certifi==2022.12.7
cffi==1.15.1
cfgrib==0.9.10.3
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.3.2
decorator==5.1.1
Django==3.2.18
django-environ==0.10.0
django-model-utils==4.3.1
djangorestframework==3.14.0
eccodes==1.5.1
exceptiongroup==1.1.1
executing==1.2.0
findlibs==0.0.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
fsspec==2023.4.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
gunicorn==20.1.0
herbie-data==2023.3.0
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
MetPy==1.4.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.20.1
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.7.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.0
pygrib==2.1.4
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
pyshp==2.3.1
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-mock==1.10.0
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data==0.6.2
strict-rfc3339==0.7
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
xarray==2023.3.0
zipp==3.15.0
